+++ py_examples/tempCodeRunnerFile.py
@@ -1,0 +1,4 @@
+x = np.logspace(0, 4, 100)  
+y1 = np.log(x)               
+y2 = np.sqrt(x)              
+y3 = x * 0.01